scripts: fix install-foundry
The latest update to the github actions runner sets the XDG_CONFIG_HOME
directory to /home/runner/.config. When this variable is set, the curl |
bash installer installs foundryup to that path. This was resulting in
foundryup being installed to /home/runner/.config/.foundry/bin/foundryup
and breaking this script.

With this change, it correctly installs to:
    /home/runner/.foundry/bin/foundryup